+++ src/engine/core/projections.ts
@@ -3,7 +3,7 @@ import {batch} from "hurdak"
 import {Worker} from "src/engine/core/utils"
 import type {Event} from "src/engine/events/model"
 import {sessions} from "src/engine/session/state"
-import {events} from "src/engine/events/state"
+import {_events} from "src/engine/events/state"
 
 export const projections = new Worker<Event>({
   getKey: prop("kind"),
@@ -15,7 +15,7 @@ projections.addGlobalHandler(
     const userEvents = chunk.filter(e => $sessions[e.pubkey])
 
     if (userEvents.length > 0) {
-      events.mapStore.update($events => {
+      _events.mapStore.update($events => {
         for (const e of userEvents) {
           $events.set(e.id, e)
         }