+++ src/engine/events/requests.ts
@@ -1,12 +1,11 @@
 import {prop} from "ramda"
-import {sumBy} from "src/util/misc"
 import {sessions} from "src/engine/session/state"
 import {getUserRelayUrls} from "src/engine/relays/utils"
 import {load} from "src/engine/network/utils"
-import {deletes} from "./state"
+import {deletesLastUpdated} from "./state"
 
 export const loadDeletes = () => {
-  const since = sumBy(prop("created_at"), deletes.get().filter(prop("created_at"))) || 0
+  const since = deletesLastUpdated.get()
   const authors = Object.values(sessions.get()).map(prop("pubkey"))
 
   return load({