+++ src/engine/events/state.ts
@@ -1,5 +1,7 @@
-import {collection} from "src/engine/core/utils"
-import type {Event, Delete} from "./model"
+import {ScalableBloomFilter} from "bloom-filters"
+import {Collection, Writable, writable} from "src/engine/core/utils"
+import type {Event} from "./model"
 
-export const events = collection<Event>("id")
-export const deletes = collection<Delete>("value")
+export const _events = new Collection<Event>("id", 1000)
+export const deletes = new Writable(new ScalableBloomFilter(), 1000)
+export const deletesLastUpdated = writable(0)