+++ src/engine/labels/index.ts
@@ -1,3 +1,4 @@
 export * from "./state"
+export * from "./derived"
 export * from "./projections"
 export * from "./commands"