+++ src/engine/labels/state.ts
@@ -1,4 +1,4 @@
 import {collection} from "src/engine/core/utils"
 import type {Event} from "src/engine/events/model"
 
-export const labels = collection<Event>("id")
+export const _labels = collection<Event>("id")