+++ src/engine/lists/state.ts
@@ -1,4 +1,4 @@
 import {collection} from "src/engine/core/utils"
 import type {List} from "./model"
 
-export const lists = collection<List>("naddr")
+export const _lists = collection<List>("naddr")