+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - [x] Fix list deletion
 - [x] Reduce repeated notes in feeds
 - [x] Add database import/export
+- [x] Use a bloom filter for deletes
 
 # 0.3.6